+++ selfdrive/ui/qt/offroad/settings.h
@@ -34,25 +34,3 @@ private:
   QButtonGroup *nav_btns;
   QStackedWidget *panel_widget;
 };
-
-class SoftwarePanel : public ListWidget {
-  Q_OBJECT
-public:
-  explicit SoftwarePanel(QWidget* parent = nullptr);
-
-private:
-  void showEvent(QShowEvent *event) override;
-  void updateLabels();
-  void checkForUpdates();
-
-  bool is_onroad = false;
-
-  QLabel *onroadLbl;
-  LabelControl *versionLbl;
-  ButtonControl *installBtn;
-  ButtonControl *downloadBtn;
-  ButtonControl *targetBranchBtn;
-
-  Params params;
-  ParamWatcher *fs_watch;
-};
