+++ selfdrive/ui/qt/offroad/software_panel.h
@@ -1,0 +1,25 @@
+#pragma once
+
+#include "selfdrive/ui/qt/offroad/settings.h"
+
+class SoftwarePanel : public ListWidget {
+  Q_OBJECT
+public:
+  explicit SoftwarePanel(QWidget* parent = nullptr);
+
+private:
+  void showEvent(QShowEvent *event) override;
+  void updateLabels();
+  void checkForUpdates();
+
+  bool is_onroad = false;
+
+  QLabel *onroadLbl;
+  LabelControl *versionLbl;
+  ButtonControl *installBtn;
+  ButtonControl *downloadBtn;
+  ButtonControl *targetBranchBtn;
+
+  Params params;
+  ParamWatcher *fs_watch;
+};
